Issue 8 - Support for YouTube
Pages can take another property youtube. The value for this property
should be the ID of a YouTube video. So for example, if you want to embed
this video:

http://www.youtube.com/watch?v=Tl1a2VN_pec

you give the page where you want to embed it this attribute:

about:
  title: "About the project"
  intro: "A short description of the project."
  youtube: "Tl1a2VN_pec"

Putting a video on a page will suppress any picture associated with the page.
(about.png in the example above.)