Issue 21 - Integrate chunks of HTML in place of pics or videos
It is now possible to fill the main content space with HTML text.
The HTML that you want to include on a given page should be placed
in a separate file, in the img directory. The name of the file should
be the slug for the page you want to display it on, and then an
".html" extension. E.g.: aboutus.html